Move semantic version logic to inventory-utils
