+++ common/inventory-utils/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod checksum;
 pub mod inv;
+pub mod semvrs;
 pub mod vrs;
